# tests/fixtures/openapi.snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "M+ Trust Factor API",
    "version": "0.1.0",
    "description": "Public character/comparison routes with SWR refresh, Battle.net OAuth IAM, and permission-protected admin routes (shared admin key retained as emergency fallback)."
  },
...
  "paths": {
...
    "/api/v1/me/characters/{ownershipId}/primary": {
      "post": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "ownershipId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ]
}